+++ api/spec/controllers/spree/api/base_controller_spec.rb
@@ -10,8 +10,9 @@ describe Spree::Api::BaseController do
 
   context "signed in as a user using an authentication extension" do
     before do
-      controller.stub :try_spree_current_user => double(:email => "spree@example.com")
-      Spree::Api::Config[:requires_authentication] = true
+      user = double(:email => "spree@example.com")
+      user.stub_chain :spree_roles, pluck: []
+      controller.stub :try_spree_current_user => user
     end
 
     it "can make a request" do
@@ -66,6 +67,7 @@ describe Spree::Api::BaseController do
 
   it 'handles exceptions' do
     subject.should_receive(:authenticate_user).and_return(true)
+    subject.should_receive(:load_user_roles).and_return(true)
     subject.should_receive(:index).and_raise(Exception.new("no joy"))
     get :index, :token => "fake_key"
     json_response.should == { "exception" => "no joy" }