+++ backend/app/controllers/spree/admin/promotions_controller.rb
@@ -12,6 +12,7 @@ module Spree
 
         def load_data
           @calculators = Rails.application.config.spree.calculators.promotion_actions_create_adjustments
+          @promotion_categories = Spree::PromotionCategory.order(:name)
         end
 
         def collection