+++ backend/spec/controllers/spree/admin/promotions_controller_spec.rb
@@ -5,11 +5,17 @@ describe Spree::Admin::PromotionsController do
 
   let!(:promotion1) { create(:promotion, name: "name1", code: "code1", path: "path1") }
   let!(:promotion2) { create(:promotion, name: "name2", code: "code2", path: "path2") }
+  let!(:category) { create :promotion_category }
 
   context "#index" do
     it "succeeds" do
       spree_get :index
       expect(assigns[:promotions]).to match_array [promotion2, promotion1]
+    end
+
+    it "assigns promotion categories" do
+      spree_get :index
+      expect(assigns[:promotion_categories]).to match_array [category]
     end
 
     context "search" do