+++ core/app/models/spree/promotion_category.rb
@@ -1,0 +1,6 @@
+module Spree
+  class PromotionCategory < Spree::Base
+    validates_presence_of :name
+    has_many :promotions
+  end
+end