+++ core/db/migrate/20140715182625_create_spree_promotion_categories.rb
@@ -1,0 +1,11 @@
+class CreateSpreePromotionCategories < ActiveRecord::Migration
+  def change
+    create_table :spree_promotion_categories do |t|
+      t.string :name
+      t.timestamps
+    end
+
+    add_column :spree_promotions, :promotion_category_id, :integer
+    add_index :spree_promotions, :promotion_category_id
+  end
+end