+++ core/db/migrate/20140716212330_add_reception_and_acceptance_status_to_return_items.rb
@@ -1,0 +1,6 @@
+class AddReceptionAndAcceptanceStatusToReturnItems < ActiveRecord::Migration
+  def change
+    add_column :spree_return_items, :reception_status, :string
+    add_column :spree_return_items, :acceptance_status, :string
+  end
+end