+++ core/spec/models/spree/order/checkout_spec.rb
@@ -47,6 +47,21 @@ describe Spree::Order do
       Spree::Order.remove_transition(nil).should be_false
     end
 
+    it "always return integer on checkout_step_index" do
+      expect(order.checkout_step_index("imnotthere")).to be_a Integer
+      expect(order.checkout_step_index("delivery")).to be > 0
+    end
+
+    it "passes delivery state when transitioning from address over delivery to payment" do
+      order.stub :payment_required? => true
+      order.state = "address"
+      order.passed_checkout_step?("delivery").should be_false
+      order.state = "delivery"
+      order.passed_checkout_step?("delivery").should be_false
+      order.state = "payment"
+      order.passed_checkout_step?("delivery").should be_true
+    end
+
     context "#checkout_steps" do
       context "when confirmation not required" do
         before do